+++ qywx-spring-boot-api/src/main/java/com/github/shuaidd/client/WeChatClient.java
@@ -50,483 +50,6 @@ public interface WeChatClient extends CommonClient {
     AccessTokenResponse getAccessToken(@RequestParam("corpid") String corpId, @RequestParam("corpsecret") String secret);
 
     /**
-     * 创建成员
-     *
-     * @param request 请求参数
-     * @param app     应用
-     * @return BaseResponse
-     */
-    @PostMapping(value = "user/create", headers = HEAD)
-    BaseResponse createUser(CreateUserRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 读取成员信息
-     *
-     * @param app    应用名
-     * @param userId 成员编号
-     * @return WeChatUserResponse
-     */
-    @GetMapping(value = "user/get", headers = HEAD)
-    WeChatUserResponse getUser(@RequestParam("userid") String userId, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 更新成员信息【根据userID更新，必传】
-     *
-     * @param app     应用名
-     * @param request 更新请求参数
-     * @return BaseResponse
-     */
-    @PostMapping(value = "user/update", headers = HEAD)
-    BaseResponse updateUser(UpdateUserRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 删除成员
-     *
-     * @param userId 成员编号
-     * @param app    应用
-     * @return BaseResponse
-     */
-    @GetMapping(value = "user/delete", headers = HEAD)
-    BaseResponse deleteUser(@RequestParam("userid") String userId, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 批量删除成员
-     *
-     * @param request 请求体
-     * @param app     应用
-     * @return BaseResponse
-     */
-    @PostMapping(value = "user/batchdelete", headers = HEAD)
-    BaseResponse batchDelete(BatchDeleteUserRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取部门成员[应用须拥有指定部门的查看权限。]
-     *
-     * @param app          应用名
-     * @param departmentId 获取的部门id
-     * @param fetchChild   1/0：是否递归获取子部门下面的成员
-     * @return DepartmentUserResponse
-     */
-    @GetMapping(value = "user/simplelist", headers = HEAD)
-    DepartmentUserResponse getDepartmentUser(@RequestParam("department_id") Integer departmentId, @RequestParam("fetch_child") Integer fetchChild, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取部门成员详情[应用须拥有指定部门的查看权限。]
-     *
-     * @param app          应用名
-     * @param departmentId 获取的部门id
-     * @param fetchChild   1/0：是否递归获取子部门下面的成员
-     * @return DepartmentUserResponse
-     */
-    @GetMapping(value = "user/list", headers = HEAD)
-    DepartmentUserResponse getDepartmentUserDetail(@RequestParam("department_id") Integer departmentId, @RequestParam("fetch_child") Integer fetchChild, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取成员ID列表
-     *
-     * @param request 请求
-     * @param app     应用名
-     * @return SimpleDeptUserResponse
-     */
-    @PostMapping(value = "user/list_id", headers = HEAD)
-    SimpleDeptUserResponse getUserIds(CursorPageRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * userid转openid
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return ConvertUserIdOpenIdResponse
-     */
-    @PostMapping(value = "user/convert_to_openid", headers = HEAD)
-    ConvertUserIdOpenIdResponse convertToOpenId(ConvertUserIdOpenIdRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * openid转userid
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return ConvertUserIdOpenIdResponse
-     */
-    @PostMapping(value = "user/convert_to_userid", headers = HEAD)
-    ConvertUserIdOpenIdResponse convertToUserId(ConvertUserIdOpenIdRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 二次验证
-     *
-     * @param userId 用户编号
-     * @param app    应用名
-     * @return BaseResponse
-     */
-    @GetMapping(value = "user/authsucc", headers = HEAD)
-    BaseResponse authSucc(@RequestParam("userid") String userId, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 邀请成员
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return BaseResponse
-     */
-    @PostMapping(value = "batch/invite", headers = HEAD)
-    BaseResponse invite(InviteUserRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 创建部门
-     *
-     * @param app     应用
-     * @param request 请求体
-     * @return CreateDepartmentResponse
-     */
-    @PostMapping(value = "department/create", headers = HEAD)
-    CreateDepartmentResponse createDepartment(DepartmentRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 更新部门
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return BaseResponse
-     */
-    @PostMapping(value = "department/update", headers = HEAD)
-    BaseResponse updateDepartment(DepartmentRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 删除部门【注：不能删除根部门；不能删除含有子部门、成员的部门】
-     *
-     * @param id  企业微信部门id
-     * @param app 应用名
-     * @return BaseResponse
-     */
-    @GetMapping(value = "department/delete", headers = HEAD)
-    BaseResponse deleteDepartment(@RequestParam(value = "id", required = false) Integer id, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 权限说明：
-     * 只能拉取token对应的应用的权限范围内的部门列表
-     * <p>
-     * 获取部门列表【获取指定部门及其下的子部门。 如果不填，默认获取全量组织架构】
-     *
-     * @param id  部门id
-     * @param app 应用名
-     * @return DepartmentListResponse
-     */
-    @GetMapping(value = "department/list", headers = HEAD)
-    DepartmentListResponse departmentList(@RequestParam(value = "id", required = false) Integer id, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取子部门ID列表
-     *
-     * @param id  部门id
-     * @param app 应用名
-     * @return SimpleDepartmentListResponse
-     */
-    @GetMapping(value = "department/simplelist", headers = HEAD)
-    SimpleDepartmentListResponse departmentSimpleList(@RequestParam(value = "id", required = false) Integer id, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 创建标签【创建的标签属于该应用，只有该应用才可以增删成员】
-     * 注意，标签总数不能超过3000个。
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return CreateTagResponse
-     */
-    @PostMapping(value = "tag/create", headers = HEAD)
-    CreateTagResponse createTag(TagRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 权限说明：
-     * 调用者必须是指定标签的创建者
-     * <p>
-     * 更新标签名字【长度限制为32个字（汉字或英文字母），标签不可与其他标签重名】
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return BaseResponse
-     */
-    @PostMapping(value = "tag/update", headers = HEAD)
-    BaseResponse updateTag(TagRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 删除标签
-     *
-     * @param id  标签ID
-     * @param app 应用名
-     * @return BaseResponse
-     */
-    @GetMapping(value = "tag/delete", headers = HEAD)
-    BaseResponse deleteTag(@RequestParam("tagid") Integer id, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取标签成员
-     *
-     * @param id  标签id
-     * @param app 应用名
-     * @return QueryTagUserResponse
-     */
-    @GetMapping(value = "tag/get", headers = HEAD)
-    QueryTagUserResponse getTagUser(@RequestParam("tagid") Integer id, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 增加标签成员
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return TagUserResponse
-     */
-    @PostMapping(value = "tag/addtagusers", headers = HEAD)
-    TagUserResponse addTagUsers(TagUserRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 删除标签成员
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return TagUserResponse
-     */
-    @PostMapping(value = "tag/deltagusers", headers = HEAD)
-    TagUserResponse deleteTagUsers(TagUserRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取标签列表
-     *
-     * @param app 应用名
-     * @return TagListResponse
-     */
-    @GetMapping(value = "tag/list", headers = HEAD)
-    TagListResponse getTagList(@RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 注意事项：
-     * <p>
-     * 模板中的部门需填写部门ID，多个部门用分号分隔，部门ID必须为数字，根部门的部门id默认为1
-     * 文件中存在、通讯录中也存在的成员，更新成员在文件中指定的字段值
-     * 文件中存在、通讯录中不存在的成员，执行添加操作
-     * 通讯录中存在、文件中不存在的成员，保持不变
-     * 成员字段更新规则：可自行添加扩展字段。文件中有指定的字段，以指定的字段值为准；文件中没指定的字段，不更新
-     * <p>
-     * <p>
-     * 增量更新成员
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return AsyncJobResponse
-     */
-    @PostMapping(value = "batch/syncuser", headers = HEAD)
-    AsyncJobResponse asyncBatchUpdateUser(AsyncJobRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 注意事项：
-     * <p>
-     * 模板中的部门需填写部门ID，多个部门用分号分隔，部门ID必须为数字，根部门的部门id默认为1
-     * 文件中存在、通讯录中也存在的成员，完全以文件为准
-     * 文件中存在、通讯录中不存在的成员，执行添加操作
-     * 通讯录中存在、文件中不存在的成员，执行删除操作。出于安全考虑，下面两种情形系统将中止导入并返回相应的错误码。
-     * 需要删除的成员多于50人，且多于现有人数的20%以上
-     * 需要删除的成员少于50人，且多于现有人数的80%以上
-     * 成员字段更新规则：可自行添加扩展字段。文件中有指定的字段，以指定的字段值为准；文件中没指定的字段，不更新
-     * <p>
-     * 全量覆盖成员
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return AsyncJobResponse
-     */
-    @PostMapping(value = "batch/replaceuser", headers = HEAD)
-    AsyncJobResponse fullCoverUser(AsyncJobRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 注意事项：
-     * <p>
-     * 文件中存在、通讯录中也存在的部门，执行修改操作
-     * 文件中存在、通讯录中不存在的部门，执行添加操作
-     * 文件中不存在、通讯录中存在的部门，当部门下没有任何成员或子部门时，执行删除操作
-     * 文件中不存在、通讯录中存在的部门，当部门下仍有成员或子部门时，暂时不会删除，当下次导入成员把人从部门移出后自动删除
-     * CSV文件中，部门名称、部门ID、父部门ID为必填字段，部门ID必须为数字，根部门的部门id默认为1；排序为可选字段，置空或填0不修改排序, order值大的排序靠前。
-     * <p>
-     * 全量覆盖部门
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return AsyncJobResponse
-     */
-    @PostMapping(value = "batch/replaceparty", headers = HEAD)
-    AsyncJobResponse fullCoverDepartment(AsyncJobRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取异步任务结果
-     *
-     * @param jobId 任务id
-     * @param app   应用名
-     * @return AsyncJobResultResponse
-     */
-    @GetMapping(value = "batch/getresult", headers = HEAD)
-    AsyncJobResultResponse jobResult(@RequestParam("jobid") String jobId, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取访问用户身份
-     *
-     * @param code 授权码
-     * @param app  应用名
-     * @return AuthenticationResponse
-     */
-    @GetMapping(value = "user/getuserinfo", headers = HEAD)
-    AuthenticationResponse getAuthentication(@RequestParam("code") String code, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取应用
-     *
-     * @param agentId 应用编号
-     * @param app     应用名
-     * @return WeChatApplicationResponse
-     */
-    @GetMapping(value = "agent/get", headers = HEAD)
-    WeChatApplicationResponse getApplication(@RequestParam("agentid") String agentId, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 设置应用
-     *
-     * @param reuqest 请求
-     * @param app     应用名
-     * @return BaseResponse
-     */
-    @PostMapping(value = "agent/set", headers = HEAD)
-    BaseResponse applicationSetting(ApplicationSettingReuqest reuqest, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 创建菜单
-     *
-     * @param request 请求体
-     * @param agentid 应用编号
-     * @param app     应用名
-     * @return BaseResponse
-     */
-    @PostMapping(value = "menu/create", headers = HEAD)
-    BaseResponse createApplicationButton(ApplicationButtonRequest request, @RequestParam("agentid") String agentid, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取菜单
-     *
-     * @param agentid 应用编号
-     * @param app     应用名
-     * @return ApplicationButtonResponse
-     */
-    @GetMapping(value = "menu/get", headers = HEAD)
-    ApplicationButtonResponse getApplicationButtons(@RequestParam("agentid") String agentid, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 删除菜单
-     *
-     * @param agentid 应用编号
-     * @param app     应用名
-     * @return BaseResponse
-     */
-    @GetMapping(value = "menu/delete", headers = HEAD)
-    BaseResponse deleteApplicationButtons(@RequestParam("agentid") String agentid, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 发送消息
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return SendMessageResponse
-     */
-    @PostMapping(value = "message/send", headers = HEAD)
-    SendMessageResponse sendMessage(SendMessageRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 创建群聊会话
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return CreateAppChatResponse
-     */
-    @PostMapping(value = "appchat/create", headers = HEAD)
-    CreateAppChatResponse createAppChat(CreateAppChatRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 修改群聊会话
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return BaseResponse
-     */
-    @PostMapping(value = "appchat/update", headers = HEAD)
-    BaseResponse updateAppChat(UpdateAppChatRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取群聊会话
-     *
-     * @param chatId 群聊编号
-     * @param app    应用名
-     * @return SearchAppChatResponse
-     */
-    @GetMapping(value = "appchat/get", headers = HEAD)
-    SearchAppChatResponse searchAppChat(@RequestParam("chatid") String chatId, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 发送群消息
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return BaseResponse
-     */
-    @PostMapping(value = "appchat/send", headers = HEAD)
-    BaseResponse sendAppChatMessage(SendAppChatRequest request, @RequestParam(HEAD_KEY) String app);
-
-
-    /**
-     * 获取打卡数据
-     * <p>
-     * 获取记录时间跨度不超过一个月
-     * 用户列表不超过100个。若用户超过100个，请分批获取
-     * 有打卡记录即可获取打卡数据，与当前”打卡应用”是否开启无关
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return CheckInDataResponse
-     */
-    @PostMapping(value = "checkin/getcheckindata", headers = HEAD)
-    CheckInDataResponse getCheckInData(CheckInDataRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取打卡规则
-     * <p>
-     * 用户列表不超过100个，若用户超过100个，请分批获取。
-     * 用户在不同日期的规则不一定相同，请按天获取。
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return CheckInRuleResponse
-     */
-    @PostMapping(value = "checkin/getcheckinoption", headers = HEAD)
-    CheckInRuleResponse getCheckInOption(CheckInRuleRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取审批数据
-     * <p>
-     * 获取审批记录请求参数endtime需要大于startime， 切起始时间跨度不要超过一个月；
-     * 一次请求返回的审批记录上限是100条，超过100条记录请使用next_spnum进行分页拉取。
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return ApprovalDataResponse
-     */
-    @PostMapping(value = "corp/getapprovaldata", headers = HEAD)
-    ApprovalDataResponse getApprovalData(ApprovalDataRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取公费电话拨打记录
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return DialRecordResponse
-     */
-    @PostMapping(value = "dial/get_dial_record", headers = HEAD)
-    DialRecordResponse getDialRecord(DialRecordRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
      * 获取手机号随机串
      *
      * @param request 请求体
@@ -535,205 +58,6 @@ public interface WeChatClient extends CommonClient {
      */
     @PostMapping(value = "user/get_mobile_hashcode", headers = HEAD)
     MobileHashCodeResponse getMobileHashcode(MobileHashCodeRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 创建预约会议
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return CreateMeetingResponse
-     */
-    @PostMapping(value = "meeting/create", headers = HEAD)
-    CreateMeetingResponse createMeeting(MeetingRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 修改预约会议
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return BaseResponse updateMeeting(MeetingRequest request, @RequestParam(HEAD_KEY) String app);
-     */
-    @PostMapping(value = "meeting/update", headers = HEAD)
-    BaseResponse updateMeeting(MeetingRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 取消预约会议
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return BaseResponse
-     */
-    @PostMapping(value = "meeting/cancel", headers = HEAD)
-    BaseResponse cancelMeeting(CancelMeetingRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取成员会议ID列表
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return GetMeetingForUserResponse
-     */
-    @PostMapping(value = "meeting/get_user_meetingid", headers = HEAD)
-    GetMeetingForUserResponse getUserMeetingId(GetMeetingForUserRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取会议详情
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return MeetingDetailResponse
-     */
-    @PostMapping(value = "meeting/get_info", headers = HEAD)
-    MeetingDetailResponse getMeetingInfo(GetMeetingDetailRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 创建预约直播
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return CreateLivingResponse
-     */
-    @PostMapping(value = "living/create", headers = HEAD)
-    CreateLivingResponse createLiving(LivingRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 修改预约直播
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return BaseResponse
-     */
-    @PostMapping(value = "living/modify", headers = HEAD)
-    BaseResponse modifyLiving(LivingRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 取消预约直播
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return BaseResponse
-     */
-    @PostMapping(value = "living/cancel", headers = HEAD)
-    BaseResponse cancelLiving(CancelLivingRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 删除直播回放
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return BaseResponse
-     */
-    @PostMapping(value = "living/delete_replay_data", headers = HEAD)
-    BaseResponse delLiving(DeleteLivingRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取微信观看直播凭证
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return LivingCodeResponse
-     */
-    @PostMapping(value = "living/get_living_code", headers = HEAD)
-    LivingCodeResponse getLivingCode(LivingCodeRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取成员直播ID列表
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return GetUserLivingResponse
-     */
-    @PostMapping(value = "living/get_user_all_livingid", headers = HEAD)
-    GetUserLivingResponse getUserAllLivingId(GetUserLivingRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取直播详情
-     *
-     * @param livingId 直播编号
-     * @param app      应用名
-     * @return LivingInfoResponse
-     */
-    @GetMapping(value = "living/get_living_info", headers = HEAD)
-    LivingInfoResponse getLivingInfo(@RequestParam("livingid") String livingId, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取直播观看明细
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return WatchStatInfoResponse
-     */
-    @PostMapping(value = "living/get_watch_stat", headers = HEAD)
-    WatchStatInfoResponse getWatchStat(GetWatchStatRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取跳转小程序商城的直播观众信息
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return LivingShareResponse
-     */
-    @PostMapping(value = "living/get_living_share_info", headers = HEAD)
-    LivingShareResponse getLivingShareInfo(LivingShareRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取企业所有打卡规则
-     *
-     * @param app 应用名
-     * @return CheckInOptionResponse
-     */
-    @PostMapping(value = "checkin/getcorpcheckinoption", headers = HEAD)
-    CheckInOptionResponse getCorpCheckInOption(@RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取打卡日报数据
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return CheckInDayReportResponse
-     */
-    @PostMapping(value = "checkin/getcheckin_daydata", headers = HEAD)
-    CheckInDayReportResponse getCheckInDayData(CommonOaRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取打卡月报数据
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return CheckInDayReportResponse
-     */
-    @PostMapping(value = "checkin/getcheckin_monthdata", headers = HEAD)
-    CheckInDayReportResponse getCheckInMonthData(CommonOaRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取打卡人员排班信息
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return CheckInScheduleResponse
-     */
-    @PostMapping(value = "checkin/getcheckinschedulist", headers = HEAD)
-    CheckInScheduleResponse getCheckInScheduList(CommonOaRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 为打卡人员排班
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return BaseResponse
-     */
-    @PostMapping(value = "checkin/setcheckinschedulist", headers = HEAD)
-    BaseResponse setCheckInScheduleList(SetCheckInScheduleRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 录入打卡人员人脸信息
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return BaseResponse
-     */
-    @PostMapping(value = "checkin/addcheckinuserface", headers = HEAD)
-    BaseResponse addCheckInUserFace(AddCheckInUserFaceRequest request, @RequestParam(HEAD_KEY) String app);
 
     /**
      * 获取企业微信API域名IP段
@@ -754,56 +78,6 @@ public interface WeChatClient extends CommonClient {
     ApiDomainIpResponse callBackIp(@RequestParam(HEAD_KEY) String app);
 
     /**
-     * 发起语音电话
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return EmergencyCallResponse
-     */
-    @PostMapping(value = "pstncc/call", headers = HEAD)
-    EmergencyCallResponse pstnccCall(EmergencyCallRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取接听状态
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return EmergencyCallResponse
-     */
-    @PostMapping(value = "pstncc/getstates", headers = HEAD)
-    GetCallStateResponse pstnccCallState(GetCallStateRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取加入企业二维码
-     *
-     * @param sizeType qrcode尺寸类型
-     * @param app      应用名
-     * @return 二维码地址
-     */
-    @GetMapping(value = "corp/get_join_qrcode", headers = HEAD)
-    JoinQrCodeResponse getJoinQrCode(@RequestParam("size_type") Integer sizeType, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 获取企业活跃成员数
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return 企业活跃成员数
-     */
-    @PostMapping(value = "user/get_active_stat", headers = HEAD)
-    ActiveStatResponse getActiveStat(ActiveStatRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 导出成员
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return 异步任务编号
-     */
-    @PostMapping(value = "export/simple_user", headers = HEAD)
-    AsyncJobResponse exportUser(AddressBookExportRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
      * 获取导出结果
      *
      * @param jobId 任务编号
@@ -812,34 +86,4 @@ public interface WeChatClient extends CommonClient {
      */
     @GetMapping(value = "export/get_result", headers = HEAD)
     ExportResultResponse getExportResult(@RequestParam("jobid") String jobId, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 导出成员详情
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return 异步任务编号
-     */
-    @PostMapping(value = "export/user", headers = HEAD)
-    AsyncJobResponse exportUserDetail(AddressBookExportRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 导出部门
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return 异步任务编号
-     */
-    @PostMapping(value = "export/department", headers = HEAD)
-    AsyncJobResponse exportDepartment(AddressBookExportRequest request, @RequestParam(HEAD_KEY) String app);
-
-    /**
-     * 导出标签成员
-     *
-     * @param request 请求体
-     * @param app     应用名
-     * @return 异步任务编号
-     */
-    @PostMapping(value = "export/taguser", headers = HEAD)
-    AsyncJobResponse exportTagUser(AddressBookExportRequest request, @RequestParam(HEAD_KEY) String app);
 }
