+++ qywx-spring-boot-api/src/main/java/com/github/shuaidd/service/EfficiencyToolService.java
@@ -1,10 +1,14 @@
 package com.github.shuaidd.service;
 
-import com.github.shuaidd.response.*;
+import com.github.shuaidd.dto.tool.DialRecord;
 import com.github.shuaidd.response.tool.*;
 import com.github.shuaidd.resquest.oa.CalendarRequest;
 import com.github.shuaidd.resquest.tool.*;
 import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
 
 /**
  * 描述 效率工具
@@ -23,7 +27,7 @@ public class EfficiencyToolService extends AbstractBaseService {
      * @return CalendarResponse
      */
     public CalendarResponse addCalendar(CalendarRequest request, String applicationName) {
-        return weChatClient.addCalendar(request, applicationName);
+        return workOaClient.addCalendar(request, applicationName);
     }
 
     /**
@@ -33,7 +37,7 @@ public class EfficiencyToolService extends AbstractBaseService {
      * @param applicationName 应用名称
      */
     public void updateCalendar(CalendarRequest request, String applicationName) {
-        weChatClient.updateCalendar(request, applicationName);
+        workOaClient.updateCalendar(request, applicationName);
     }
 
     /**
@@ -44,7 +48,7 @@ public class EfficiencyToolService extends AbstractBaseService {
      * @return GetCalendarResponse
      */
     public GetCalendarResponse getCalendar(GetCalendarRequest request, String applicationName) {
-        return weChatClient.getCalendar(request, applicationName);
+        return workOaClient.getCalendar(request, applicationName);
     }
 
     /**
@@ -54,7 +58,7 @@ public class EfficiencyToolService extends AbstractBaseService {
      * @param applicationName 应用名称
      */
     public void delCalendar(DeleteCalendarRequest request, String applicationName) {
-        weChatClient.delCalendar(request, applicationName);
+        workOaClient.delCalendar(request, applicationName);
     }
 
     /**
@@ -65,7 +69,7 @@ public class EfficiencyToolService extends AbstractBaseService {
      * @return AddScheduleResponse
      */
     public AddScheduleResponse addSchedule(ScheduleRequest request, String applicationName) {
-        return weChatClient.addSchedule(request, applicationName);
+        return workOaClient.addSchedule(request, applicationName);
     }
 
     /**
@@ -75,7 +79,7 @@ public class EfficiencyToolService extends AbstractBaseService {
      * @param applicationName 应用名称
      */
     public void updateSchedule(ScheduleRequest request, String applicationName) {
-        weChatClient.updateSchedule(request, applicationName);
+        workOaClient.updateSchedule(request, applicationName);
     }
 
     /**
@@ -86,7 +90,7 @@ public class EfficiencyToolService extends AbstractBaseService {
      * @return GetScheduleResponse
      */
     public GetScheduleResponse getSchedules(GetScheduleRequest request, String applicationName) {
-        return weChatClient.getSchedules(request, applicationName);
+        return workOaClient.getSchedules(request, applicationName);
     }
 
     /**
@@ -96,7 +100,7 @@ public class EfficiencyToolService extends AbstractBaseService {
      * @param applicationName 应用名称
      */
     public void delSchedule(CancelScheduleRequest request, String applicationName) {
-        weChatClient.delSchedule(request, applicationName);
+        workOaClient.delSchedule(request, applicationName);
     }
 
     /**
@@ -107,7 +111,7 @@ public class EfficiencyToolService extends AbstractBaseService {
      * @return GetScheduleResponse
      */
     public GetScheduleResponse getScheduleByCalendar(GetScheduleByCalendarRequest request, String applicationName) {
-        return weChatClient.getScheduleByCalendar(request, applicationName);
+        return workOaClient.getScheduleByCalendar(request, applicationName);
     }
 
     /**
@@ -118,7 +122,7 @@ public class EfficiencyToolService extends AbstractBaseService {
      * @return CreateMeetingResponse
      */
     CreateMeetingResponse createMeeting(MeetingRequest request, String applicationName) {
-        return weChatClient.createMeeting(request, applicationName);
+        return efficiencyToolClient.createMeeting(request, applicationName);
     }
 
     /**
@@ -128,7 +132,7 @@ public class EfficiencyToolService extends AbstractBaseService {
      * @param applicationName 应用名称
      */
     public void updateMeeting(MeetingRequest request, String applicationName) {
-        weChatClient.updateMeeting(request, applicationName);
+        efficiencyToolClient.updateMeeting(request, applicationName);
     }
 
     /**
@@ -138,7 +142,7 @@ public class EfficiencyToolService extends AbstractBaseService {
      * @param applicationName 应用名称
      */
     public void cancelMeeting(CancelMeetingRequest request, String applicationName) {
-        weChatClient.cancelMeeting(request, applicationName);
+        efficiencyToolClient.cancelMeeting(request, applicationName);
     }
 
     /**
@@ -149,7 +153,7 @@ public class EfficiencyToolService extends AbstractBaseService {
      * @return GetMeetingForUserResponse
      */
     public GetMeetingForUserResponse getUserMeetingId(GetMeetingForUserRequest request, String applicationName) {
-        return weChatClient.getUserMeetingId(request, applicationName);
+        return efficiencyToolClient.getUserMeetingId(request, applicationName);
     }
 
     /**
@@ -160,7 +164,7 @@ public class EfficiencyToolService extends AbstractBaseService {
      * @return MeetingDetailResponse
      */
     public MeetingDetailResponse getUserMeetingInfo(GetMeetingDetailRequest request, String applicationName) {
-        return weChatClient.getMeetingInfo(request, applicationName);
+        return efficiencyToolClient.getMeetingInfo(request, applicationName);
     }
 
     /**
@@ -171,7 +175,7 @@ public class EfficiencyToolService extends AbstractBaseService {
      * @return CreateLivingResponse
      */
     public CreateLivingResponse createLiving(LivingRequest request, String applicationName) {
-        return weChatClient.createLiving(request, applicationName);
+        return efficiencyToolClient.createLiving(request, applicationName);
     }
 
     /**
@@ -181,7 +185,7 @@ public class EfficiencyToolService extends AbstractBaseService {
      * @param applicationName 应用名
      */
     public void modifyLiving(LivingRequest request, String applicationName) {
-        weChatClient.modifyLiving(request, applicationName);
+        efficiencyToolClient.modifyLiving(request, applicationName);
     }
 
     /**
@@ -191,7 +195,7 @@ public class EfficiencyToolService extends AbstractBaseService {
      * @param applicationName 应用名
      */
     public void cancelLiving(CancelLivingRequest request, String applicationName) {
-        weChatClient.cancelLiving(request, applicationName);
+        efficiencyToolClient.cancelLiving(request, applicationName);
     }
 
     /**
@@ -201,7 +205,7 @@ public class EfficiencyToolService extends AbstractBaseService {
      * @param applicationName 应用名
      */
     public void delLiving(DeleteLivingRequest request, String applicationName) {
-        weChatClient.delLiving(request, applicationName);
+        efficiencyToolClient.delLiving(request, applicationName);
     }
 
     /**
@@ -212,7 +216,7 @@ public class EfficiencyToolService extends AbstractBaseService {
      * @return LivingCodeResponse
      */
     public LivingCodeResponse getLivingCode(LivingCodeRequest request, String applicationName) {
-        return weChatClient.getLivingCode(request, applicationName);
+        return efficiencyToolClient.getLivingCode(request, applicationName);
     }
 
     /**
@@ -223,7 +227,7 @@ public class EfficiencyToolService extends AbstractBaseService {
      * @return GetUserLivingResponse
      */
     public GetUserLivingResponse getUserAllLivingId(GetUserLivingRequest request, String applicationName) {
-        return weChatClient.getUserAllLivingId(request, applicationName);
+        return efficiencyToolClient.getUserAllLivingId(request, applicationName);
     }
 
     /**
@@ -234,7 +238,7 @@ public class EfficiencyToolService extends AbstractBaseService {
      * @return LivingInfoResponse
      */
     public LivingInfoResponse getLivingInfo(String livingId, String applicationName) {
-        return weChatClient.getLivingInfo(livingId, applicationName);
+        return efficiencyToolClient.getLivingInfo(livingId, applicationName);
     }
 
     /**
@@ -245,7 +249,7 @@ public class EfficiencyToolService extends AbstractBaseService {
      * @return WatchStatInfoResponse
      */
     public WatchStatInfoResponse getWatchStat(GetWatchStatRequest request, String applicationName) {
-        return weChatClient.getWatchStat(request, applicationName);
+        return efficiencyToolClient.getWatchStat(request, applicationName);
     }
 
     /**
@@ -256,6 +260,18 @@ public class EfficiencyToolService extends AbstractBaseService {
      * @return LivingShareResponse
      */
     public LivingShareResponse getLivingShareInfo(LivingShareRequest request, String applicationName) {
-        return weChatClient.getLivingShareInfo(request, applicationName);
+        return efficiencyToolClient.getLivingShareInfo(request, applicationName);
+    }
+
+    /**
+     * 获取公费电话拨打记录
+     *
+     * @param request         请求
+     * @param applicationName 应用名称
+     * @return DialRecord
+     */
+    public final List<DialRecord> getDialRecord(DialRecordRequest request, String applicationName) {
+        Objects.requireNonNull(request, "参数为空");
+        return efficiencyToolClient.getDialRecord(request, applicationName).getRecords();
     }
 }
